Settings: Move force dark mode to dark mode settings
