+++ res/xml/dark_mode_settings.xml
@@ -86,4 +86,9 @@
         android:summary="@string/berry_black_theme_summary"
         android:defaultValue="false"
         settings:requiresPackage="com.android.system.theme.black" />
+
+    <SwitchPreferenceCompat
+        android:key="hwui_force_dark"
+        android:title="@string/hwui_force_dark_title"
+        android:summary="@string/hwui_force_dark_summary" />
 </PreferenceScreen>
