+++ res/xml/development_settings.xml
@@ -620,11 +620,6 @@
             android:entryValues="@array/show_non_rect_clip_values" />
 
         <SwitchPreferenceCompat
-            android:key="hwui_force_dark"
-            android:title="@string/hwui_force_dark_title"
-            android:summary="@string/hwui_force_dark_summary" />
-
-        <SwitchPreferenceCompat
             android:key="force_msaa"
             android:title="@string/force_msaa"
             android:summary="@string/force_msaa_summary" />
