+++ src/com/android/settings/display/darkmode/DarkModeSettingsFragment.java
@@ -75,6 +75,7 @@ public class DarkModeSettingsFragment extends DashboardFragment {
                 DARK_THEME_END_TIME, this);
         controllers.add(mCustomStartController);
         controllers.add(mCustomEndController);
+        controllers.add(new ForceDarkPreferenceController(context));
         return controllers;
     }
